v1.9 alpha reflect

+++ src/main/java/com/minecraft/moonlake/util/StringUtil.java
@@ -568,7 +568,7 @@ public class StringUtil {
      * @param str 字符串
      * @return 单精度浮点数, 异常返回 0
      */
-    public static double parseFloat(String str) {
+    public static float parseFloat(String str) {
 
         return parseFloat(str, 0f);
     }
@@ -580,7 +580,7 @@ public class StringUtil {
      * @param def 默认值
      * @return 单精度浮点数, 异常返回默认值
      */
-    public static double parseFloat(String str, float def) {
+    public static float parseFloat(String str, float def) {
 
         return parseFloat((Object) str, def);
     }
@@ -591,7 +591,7 @@ public class StringUtil {
      * @param obj 对象
      * @return 单精度浮点数, 异常返回 0
      */
-    public static double parseFloat(Object obj) {
+    public static float parseFloat(Object obj) {
 
         return parseFloat(obj, 0f);
     }
@@ -603,7 +603,7 @@ public class StringUtil {
      * @param def 默认值
      * @return 单精度浮点数, 异常返回默认值
      */
-    public static double parseFloat(Object obj, float def) {
+    public static float parseFloat(Object obj, float def) {
 
         if(obj instanceof Number)
             return ((Number) obj).floatValue();
